Improve handling of local variables not replaced by register equivalents in generalized constant propagation (to more reliably get warnings on out-of bounds array accesses and dereferences of invalid pointers).
git-svn-id: https://svn.code.sf.net/p/sdcc/code/branches/arraybounds/sdcc@15862 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ src/SDCCopt.c
@@ -2194,8 +2194,8 @@ checkStaticArrayParams (ebbIndex *ebbi)
           {
             const struct valinfo v = getOperandValinfo (ic, ic->left);
             wassert (IS_OP_LITERAL (ic->right));
-            unsigned long roff = operandLitValue (ic->right);
-            if (roff >= v.maxsize)
+            long long roff = operandLitValue (ic->right);
+            if (roff >= (long long)v.maxsize)
               werrorfl (ic->filename, ic->lineno, W_INVALID_PTR_DEREF);
           }
       }
